refactor: move kafka dataset to separate file
